+++ samples/shared/src/commonMain/kotlin/eric/bitria/samples/Navigation.kt
@@ -9,4 +9,5 @@ sealed class Screen {
     data object Solar : Screen()
     data object Animation : Screen()
     data object SplitView : Screen()
+    data object Environment : Screen()
 }